requirements.txt
# Data
pandas
numpy
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Features
pandas-ta==0.3.14b0

# ML
scikit-learn==1.4.0
xgboost==2.0.3

# Visualization
matplotlib
seaborn

# Utils
python-dotenv
requests==2.31.0
apscheduler
pyarrow
joblib
